api_requirements.txt
captum==0.6.0
datasets==2.11.0
fastapi==0.88.0
lightning==2.0.2
matplotlib==3.7.1
numpy==1.24.1
Pillow==9.5.0
pyyaml==6.0
scikit_learn==1.2.2
torch==2.0.0
torchmetrics==0.11.4
torchvision==0.15.1
transformers==4.28.1
uvicorn==0.21.1
